requirements.txt
numpy>=1.26.4
numba>=0.59.1
opensimplex>=0.4.5.1
pygame_menu
pygame-ce
moderngl>=5.10.0
PyGLM
opensimplex[numpy]
moderngl
glm>=0.4.4
screeninfo